move to uint8

+++ src/main.cpp
@@ -47,11 +47,11 @@ bool shouldSaveConfig = false;
 WiFiClient esp_wifi_client;
 PubSubClient mqtt_client(esp_wifi_client);
 
-int garage_door_1_state = 0;
-int garage_door_2_state = 0;
+uint8_t garage_door_1_state = 0;
+uint8_t garage_door_2_state = 0;
 
-int garage_door_1_ideal_state = 0;
-int garage_door_2_ideal_state = 0;
+uint8_t garage_door_1_ideal_state = 0;
+uint8_t garage_door_2_ideal_state = 0;
 
 unsigned long garage_door_1_last_command_millis = 0;
 unsigned long garage_door_2_last_command_millis = 0;
@@ -248,21 +248,21 @@ void loop() {
   resetRelay(garage_door_2_state,millis_since_garage_door_2_command,GARAGE_DOOR_2_COMMAND_PIN);
   mqtt_loop();
   //sensors return 1 if open 0 if closed
-  int current_garage_door_1_state = digitalRead(GARAGE_DOOR_1_SENDOR_PIN);
-  int current_garage_door_2_state = digitalRead(GARAGE_DOOR_2_SENDOR_PIN);
-  int new_garage_door_1_state = getNewGarageDoorState(garage_door_1_state, current_garage_door_1_state, millis_since_garage_door_1_command);
+  uint8_t current_garage_door_1_state = digitalRead(GARAGE_DOOR_1_SENDOR_PIN);
+  uint8_t current_garage_door_2_state = digitalRead(GARAGE_DOOR_2_SENDOR_PIN);
+  uint8_t new_garage_door_1_state = getNewGarageDoorState(garage_door_1_state, current_garage_door_1_state, millis_since_garage_door_1_command);
   if(garage_door_1_state != new_garage_door_1_state){
     mqtt_client.publish(garage_door_1_state_topic, String(garage_door_1_state).c_str());
     garage_door_1_state = new_garage_door_1_state;
   }
-  int new_garage_door_2_state = getNewGarageDoorState(garage_door_2_state, current_garage_door_2_state, millis_since_garage_door_2_command);
+  uint8_t new_garage_door_2_state = getNewGarageDoorState(garage_door_2_state, current_garage_door_2_state, millis_since_garage_door_2_command);
   if(garage_door_2_state != new_garage_door_2_state){
     mqtt_client.publish(garage_door_2_state_topic, String(current_garage_door_2_state).c_str());
     garage_door_2_state = new_garage_door_2_state;
   }
 }
 
-void resetRelay(int garage_door_state, unsigned long millis_since_last_command, uint8_t pin){
+void resetRelay(uint8_t garage_door_state, unsigned long millis_since_last_command, uint8_t pin){
   bool is_door_opening_closing = (garage_door_state == GARAGE_DOOR_OPENING_STATE || garage_door_state == GARAGE_DOOR_CLOSING_STATE);
   bool is_relay_activation_in_progress = millis_since_last_command<RELAY_ACTIVATION_MILLIS;
   if(is_door_opening_closing && !is_relay_activation_in_progress){
@@ -270,7 +270,7 @@ void resetRelay(int garage_door_state, unsigned long millis_since_last_command, 
   }
 }
 
-int getNewGarageDoorState(int previous_garage_door_state, int current_sensor_value, unsigned long millis_since_last_command){
+uint8_t getNewGarageDoorState(uint8_t previous_garage_door_state, uint8_t current_sensor_value, unsigned long millis_since_last_command){
   if(previous_garage_door_state != current_sensor_value){
     bool isGarageDoorStillOpening = 
       previous_garage_door_state == GARAGE_DOOR_OPENING_STATE &&
@@ -293,7 +293,7 @@ void mqtt_callback(char* topic, byte* payload, unsigned int length) {
   Serial.println("] ");
   
   if(length == 1 && (payload[0] == '0' || payload[0] == '1')){
-    int new_ideal_state = '0'-payload[0];
+    uint8_t new_ideal_state = '0'-payload[0];
     if(strcmp(topic, garage_door_1_command_topic) == 0) {
       garage_door_1_state = checkCommandForNewActions(new_ideal_state,garage_door_1_state,GARAGE_DOOR_1_COMMAND_PIN);
     }
